requirements/development.txt
aniso8601==7.0.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
dateparser==1.1.8
defusedxml==0.7.1
Django==3.2.18
django-environ==0.10.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
oauthlib==3.2.2
Pillow==9.5.0
promise==2.3
prompt-toolkit==2.0.9
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
r-django-jet==2.0.9
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
Rx==1.6.3
six==1.16.0
social-auth-core==4.4.2
SQLAlchemy==1.4.48
sqlparse==0.4.4
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.3
wcwidth==0.2.6
zipp==3.15.0
